LR 7 completed

+++ java_domination_339-java_labs_2024/src/main/java/ru/ssau/tk/java_domination_339/java_labs_2024/functions/MathFunction.java
@@ -7,7 +7,7 @@ public interface MathFunction {
         return this.getClass().getSimpleName();
     }
 
-    default int HashName() {
+    default long HashName() {
         return this.Name().hashCode();
     }
